// src/shared/components/treatment-requests-service-form/data-access/interfaces/treatment-requests-service-form.interface.ts
export interface TreatmentRequestsServiceDetailsInterface {
    numero_dossier: string | null;
    confirmation_contrat: boolean;
    id: number | null;
    numero_demande: string | null;
    description: string | null;
    justificatif: string | null;
    operation: string | null;
    cloture_a: string | null;
    nb_demande_soumises: number | null;
    nb_demande_traitees: number | null;
    nb_demande_cloturees: number | null;
    initie_par: number | null;
    recu_paiement: string | null;
    intervenant_id: number | null;
    traite_a: string | null;
    traite_par: number | null;
    cloture_par: number | null;
    abandonne_a: string | null;
    abandonne_par: number | null;
    acquitte_a: string | null;
    acquitte_par: number | null;
    commentaire_traitement: string | null;
    commentaire_cloture: string | null;
    nb_demande_identifiees: string | null;
    nb_demande_echouees: string | null;
    type: string | null;
    finalise_a: string | null;
    finalise_par: number | null;
    commentaire_finalisation: string | null;
    montant_formule: string | null;
    nb_cycle: number | null;
    nb_demande_crees: number | null;
    accepte: string | null;
    equipe_id: number | null;
    type_paiement: string | null;
    nb_sim_affectees: number | null;
    accepte_approbation: string | null;
    commentaire_approbation: string | null;
    approuve_par: number | null;
    approuve_a: string | null;
    prix_unitaire: string | null;
    prix_ht: string | null;
    tva: string | null;
    prix_tva: string | null;
    prix_ttc: string | null;
    delai_ack: number | null;
    delai_traitement: number | null;
    delai_cloture: number | null;
    delai_escalade: number | null;
    source: string | null;
    date_soumission_facture: string | null;
    etat_paiement: string | null;
    recuperer_par: number | null;
    recuperer_a: string | null;
    commentaire_livraison: string | null;
    etat_soumission: string | null;
    etat_traitement: string | null;
    etat_finalisation: string | null;
    etat_cloture: string | null;
    statut: string | null;
    traitement: string | null;
    notation_cloture: string | null;
    created_at: string | null;
    updated_at: string | null;
    demandeur_nom: string | null;
    demandeur_prenoms: string | null;
    demandeur_contacts: string | null;
    demandeur_email: string | null;
    demandeur_matricule: string | null;
    fournisseurInfo: FournisseurInfoInterface | null;
    clientInfo: ClientInfoInterface | null;
    prix_ttc_lettres: string | null;
    montant_total: string | null;
    impaye_anterieur: string | null;
    soldeDisponible: string | null;
    SoldeEngage: string | null;
    intervenant: any | null;
    finalisateur: any | null;
    approbateur: any | null;
    livreur: any | null;
    tenant: TenantInfo;
    facture: any | null;
    type_form: string | null;
    qte: number | null;
}
// ...
export interface FournisseurInfoInterface {
    nom_entreprise: string | null;
    forme_juridique: string | null;
    pays: string | null;
    nom_responsable: string | null;
    contact_responsable: string | null;
    email_responsable: string | null;
    rccm: string | null;
    cc: string | null;
    capital_social: string | null;
    adresse: string | null;
    email: string | null;
    contact_1: string | null;
    contact_2: string | null;
    ordre: string | null;
    banque: string | null;
    compte: string | null;
    logo: string | null;
}

export interface ClientInfoInterface {
    adresse: string | null;
    compte_client: string | null;
    nom_gerant: string | null;
    contact_admin_tenant: string | null;
    email_admin_client: string | null;
    nom_tenant: string | null;
    segment_client: string | null;
}

export interface TenantInfo {
    id: number | null;
    transaction: string | null;
    numero_demande: string | null;
    tenant_code: string | null;
    nom_client: string;
    logo_tenant: string | null;
    slug: string | null;
    domaine_activite: string;
    email_diffusion: string | null;
    url_detection_appro: string | null;
    nom_admin_tenant: string | null;
    contact_admin_tenant: string | null;
    email_admin_client: string | null;
    nom_gestionnaire: string | null;
    contact_gestionnnaire: string | null;
    email_gestionnnaire: string | null;
    gestionnaire_tenant_id: number | null;
    escalade_tenant_id: number | null;
    nom_commercial: string | null;
    contact_commercial: string | null;
    email_commercial: string | null;
    nom_senior_manager: string | null;
    contact_senior_manager: string | null;
    email_senior_manager: string | null;
    description: string | null;
    url_frontend: string | null;
    suffixe_email: string | null;
    compte_client: string;
    segment_client: string | null;
    forme_juridique_code: string;
    regime_code: string;
    adresse: string;
    numero_rccm: string;
    fichier_rccm: string | null;
    fichier_dfe: string | null;
    numero_cc: string;
    centre: string;
    nom_gerant: string;
    contact_gerant: string;
    email_gerant: string;
    piece_gerant: string | null;
    statut: string | null;
    created_at: string | null;
    updated_at: string | null;
    nb_demandes: number;
    type_entreprise: string;
}